+++ src/components/EditSections/EditUserInfo/components/ProfilePicture/ProfilePicture.js
@@ -8,6 +8,7 @@ import {
   DropdownMenu,
   Icon,
   Label,
+  Loading,
 } from '@folio/stripes/components';
 import { useStripes } from '@folio/stripes/core';
 
@@ -35,14 +36,14 @@ const ProfilePicture = ({ profilePictureId, form, personal }) => {
     const profilePictureSrc = isProfilePictureLinkAURL ? profilePictureLink : 'data:;base64,' + profilePictureData;
     const imgSrc = isFetching || !hasProfilePicture ? profilePicThumbnail : profilePictureSrc;
 
-    return (
+    return isFetching ? 
+      <span data-testid='profile-picture-loader'> <Loading/> </span> :
       <img
         data-testid="profile-picture"
         className={css.profilePlaceholder}
         alt={intl.formatMessage({ id: 'ui-users.information.profilePicture' })}
         src={imgSrc}
       />
-    );
   };
 
   const toggleExternalLinkModal = () => {